Update auth-context to include user information from token and redirect based on user role
The auth-context file has been updated to include the user information from the token in the setUser function. Additionally, the code now redirects the user to the appropriate page based on their role. If the user has the role of "ADMIN", they are redirected to the "/dashboard/products" page. Otherwise, they are redirected to the home page ("/"). This improves the user experience by providing personalized navigation after successful login.